+++ common/components/AddressField.tsx
@@ -1,6 +1,8 @@
 import React from 'react';
 import { AddressFieldFactory } from './AddressFieldFactory';
 import { donationAddressMap } from 'config';
+import translate from 'translations';
+import { Input } from 'components/ui';
 
 interface Props {
   isReadOnly?: boolean;
@@ -10,14 +12,20 @@ export const AddressField: React.SFC<Props> = ({ isReadOnly }) => (
   <AddressFieldFactory
     withProps={({ currentTo, isValid, onChange, readOnly }) => (
       <React.Fragment>
-        <input
-          className={`form-control ${isValid ? 'is-valid' : 'is-invalid'}`}
-          type="text"
-          value={currentTo.raw}
-          placeholder={donationAddressMap.ETH}
-          readOnly={!!(isReadOnly || readOnly)}
-          onChange={onChange}
-        />
+        <div className="input-group-wrapper">
+          <label className="input-group">
+            <div className="input-group-header">{translate('SEND_addr')}</div>
+            <Input
+              className={`input-group-input ${isValid ? '' : 'invalid'}`}
+              type="text"
+              value={currentTo.raw}
+              placeholder={donationAddressMap.ETH}
+              readOnly={!!(isReadOnly || readOnly)}
+              spellCheck={false}
+              onChange={onChange}
+            />
+          </label>
+        </div>
       </React.Fragment>
     )}
   />
