+++ common/components/UnitDropDown/UnitDropDown.tsx
@@ -1,7 +1,6 @@
 import React, { Component } from 'react';
 import { setUnitMeta, TSetUnitMeta } from 'actions/transaction';
 import Dropdown from 'components/ui/Dropdown';
-import { withConditional } from 'components/hocs';
 import { TokenBalance, MergedToken, getShownTokenBalances, getTokens } from 'selectors/wallet';
 import { Query } from 'components/renderCbs';
 import { connect } from 'react-redux';
@@ -22,34 +21,28 @@ interface StateProps {
   network: NetworkConfig;
 }
 
-const StringDropdown = Dropdown as new () => Dropdown<string>;
-const ConditionalStringDropDown = withConditional(StringDropdown);
-
 class UnitDropdownClass extends Component<DispatchProps & StateProps> {
   public render() {
     const { tokens, allTokens, showAllTokens, unit, network } = this.props;
     const focusedTokens = showAllTokens ? allTokens : tokens;
     return (
-      <div className="input-group-btn">
-        <Query
-          params={['readOnly']}
-          withQuery={({ readOnly }) => (
-            <ConditionalStringDropDown
-              options={[network.unit, ...getTokenSymbols(focusedTokens)]}
-              value={unit === 'ether' ? network.unit : unit}
-              condition={!readOnly}
-              conditionalProps={{
-                onChange: this.handleOnChange
-              }}
-              ariaLabel={'dropdown'}
-            />
-          )}
-        />
-      </div>
+      <Query
+        params={['readOnly']}
+        withQuery={({ readOnly }) => (
+          <Dropdown
+            options={[network.unit, ...getTokenSymbols(focusedTokens)]}
+            value={unit === 'ether' ? network.unit : unit}
+            onChange={this.handleOnChange}
+            searchable={false}
+            clearable={false}
+            disabled={!!readOnly}
+          />
+        )}
+      />
     );
   }
-  private handleOnChange = (unit: string) => {
-    this.props.setUnitMeta(unit);
+  private handleOnChange = unit => {
+    this.props.setUnitMeta(unit.value);
   };
 }
 const getTokenSymbols = (tokens: (TokenBalance | MergedToken)[]) => tokens.map(t => t.symbol);
