+++ common/components/ui/Dropdown.tsx
@@ -1,123 +1,54 @@
-import React, { PureComponent } from 'react';
-import classnames from 'classnames';
-import DropdownShell from './DropdownShell';
+import React from 'react';
+import Select, { ReactSelectProps } from 'react-select';
 
-interface Props<T> {
-  value: T | undefined;
-  options: T[];
-  ariaLabel: string;
-  label?: string;
-  extra?: any;
-  size?: string;
-  color?: string;
-  menuAlign?: string;
-  formatTitle?(option: T): any;
-  onChange?(value: T): void;
+interface Props {
+  className?: string;
+  options: any;
+  onChange: any;
 }
 
-interface State {
-  search: string;
-}
-
-export default class DropdownComponent<T> extends PureComponent<Props<T>, State> {
+export default class Dropdown extends React.Component<ReactSelectProps & Props> {
   public state = {
-    search: ''
+    selectedOption: { value: '', label: '' }
   };
 
-  private dropdownShell: DropdownShell | null;
+  public handleChange = selectedOption => {
+    this.setState({ selectedOption });
+  };
+
+  public formatOptions = options => {
+    if (typeof options[0] === 'object') {
+      return options;
+    }
+    const formatted = options.map(opt => {
+      return { value: opt, label: opt };
+    });
+    return formatted;
+  };
 
   public render() {
-    const { ariaLabel, color, size } = this.props;
+    const { onChange } = this.props;
+    const { selectedOption } = this.state;
+    const value = selectedOption && selectedOption.value;
 
     return (
-      <DropdownShell
-        renderLabel={this.renderLabel}
-        renderOptions={this.renderOptions}
-        size={size}
-        color={color}
-        ariaLabel={ariaLabel}
-        ref={el => (this.dropdownShell = el)}
+      <Select
+        // use ref to prevent <label /> from stealing focus when used inline with an input
+        ref={el => {
+          if (!!el && !!(el as any).control) {
+            (el as any).control.addEventListener('click', e => {
+              e.preventDefault();
+            });
+          }
+        }}
+        value={value}
+        onChange={obj => {
+          this.handleChange(obj);
+          onChange();
+        }}
+        {...this.props}
+        options={this.formatOptions(this.props.options)}
       />
     );
   }
-
-  private renderLabel = () => {
-    const { value } = this.props;
-    const labelStr = this.props.label ? `${this.props.label}:` : '';
-    return (
-      <span>
-        {labelStr} {this.formatTitle(value)}
-      </span>
-    );
-  };
-
-  private renderOptions = () => {
-    const { options, value, menuAlign, extra } = this.props;
-    const { search } = this.state;
-    const searchable = options.length > 20;
-    const menuClass = classnames({
-      'dropdown-menu': true,
-      [`dropdown-menu-${menuAlign || ''}`]: !!menuAlign
-    });
-    const searchableStyle = {
-      maxHeight: '300px',
-      overflowY: 'auto'
-    };
-    const searchRegex = new RegExp(search, 'gi');
-    const onSearchChange = e => {
-      this.setState({ search: e.target.value });
-    };
-
-    return (
-      <ul className={menuClass} style={searchable ? searchableStyle : {}}>
-        {searchable && (
-          <input
-            className="form-control"
-            placeholder={'Search'}
-            onChange={onSearchChange}
-            value={search}
-          />
-        )}
-
-        {options
-          .filter(option => {
-            if (searchable && search.length) {
-              return option.toString().match(searchRegex);
-            }
-            return true;
-          })
-          .map((option, i) => {
-            return (
-              <li key={i}>
-                <a
-                  className={option === value ? 'active' : ''}
-                  onClick={this.onChange.bind(null, option)}
-                >
-                  {this.props.formatTitle ? this.formatTitle(option) : option}
-                </a>
-              </li>
-            );
-          })}
-        {extra && <li key={'separator'} role="separator" className="divider" />}
-        {extra}
-      </ul>
-    );
-  };
-
-  private formatTitle = (option: any) => {
-    if (this.props.formatTitle) {
-      return this.props.formatTitle(option);
-    } else {
-      return option;
-    }
-  };
-
-  private onChange = (value: any) => {
-    if (this.props.onChange) {
-      this.props.onChange(value);
-    }
-    if (this.dropdownShell) {
-      this.dropdownShell.close();
-    }
-  };
 }
