+++ common/components/ui/Input.scss
@@ -1,0 +1,60 @@
+.example-form {
+  max-width: 744px;
+  margin: auto;
+}
+
+.input-group {
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  margin-bottom: 1rem;
+  &-header > label,
+  &-header {
+    font-size: 1rem;
+    margin-bottom: 8px;
+    font-weight: 400;
+  }
+  &-input {
+    width: 100%;
+    border: 1px solid #e5ecf3;
+    border-radius: 2px;
+    padding: 0.75rem 1rem;
+    font-weight: 400;
+    font-size: 1rem;
+    color: rgba(0, 0, 0, 0.87);
+    box-shadow: inset 0 1px 0 0 rgba(63, 63, 68, 0.05);
+    transition: border-color 120ms, box-shadow 120ms;
+    &.invalid.has-blurred {
+      border-color: #ec4d3c;
+      box-shadow: inset 0px 0px 0px 1px #ec4d3c;
+    }
+    &::placeholder {
+      color: rgba(0, 0, 0, 0.3);
+    }
+    &:focus {
+      border-color: #4295bc;
+      box-shadow: inset 0px 0px 0px 1px #4295bc;
+      &.valid {
+        border-color: #8dd17b;
+        box-shadow: inset 0px 0px 0px 1px #8dd17b;
+      }
+    }
+  }
+}
+
+.input-group-inline-dropdown {
+  display: flex;
+  flex-direction: row;
+  font-size: 1rem;
+  flex-wrap: wrap;
+  > .input-group-header {
+    width: 100%;
+  }
+  > .input-group-input {
+    flex-grow: 1;
+    width: auto;
+  }
+  > .Select {
+    margin-left: 8px;
+  }
+}
