+++ common/components/ui/index.ts
@@ -1,5 +1,6 @@
 export { default as ColorDropdown } from './ColorDropdown';
-export { default as Dropdown } from './Dropdown';
+export { default as OldDropDown } from './OldDropdown';
+export { default as DropDown } from './Dropdown';
 export { default as DropdownShell } from './DropdownShell';
 export { default as Identicon } from './Identicon';
 export { default as Modal } from './Modal';
@@ -12,6 +13,7 @@ export { default as SwapDropdown } from './SwapDropdown';
 export { default as Tooltip } from './Tooltip';
 export { default as TitleBar } from './TitleBar';
 export { default as HelpLink } from './HelpLink';
+export { default as Input } from './Input';
 export * from './ConditionalInput';
 export * from './Expandable';
 export * from './InlineSpinner';
